refactor: move semantic annotations from prothor_semantic_annotations.py to semantic_annotations.py
	modified:   src/semantic_digital_twin/adapters/procthor/procthor_semantic_annotations.py
	modified:   src/semantic_digital_twin/orm/ormatic_interface.py
	modified:   src/semantic_digital_twin/semantic_annotations/semantic_annotations.py
	modified:   test/test_procthor/test_procthor.py
