+++ src/semantic_digital_twin/orm/ormatic_interface.py
@@ -1717,7 +1717,7 @@ class HasBodyDAO(
 class BaseballDAO(
     HasBodyDAO,
     DataAccessObject[
-        semantic_digital_twin.adapters.procthor.procthor_semantic_annotations.Baseball
+        semantic_digital_twin.semantic_annotations.semantic_annotations.Baseball
     ],
 ):
 
@@ -1736,7 +1736,7 @@ class BaseballDAO(
 class BedDAO(
     HasBodyDAO,
     DataAccessObject[
-        semantic_digital_twin.adapters.procthor.procthor_semantic_annotations.Bed
+        semantic_digital_twin.semantic_annotations.semantic_annotations.Bed
     ],
 ):
 
@@ -1755,7 +1755,7 @@ class BedDAO(
 class BookDAO(
     HasBodyDAO,
     DataAccessObject[
-        semantic_digital_twin.adapters.procthor.procthor_semantic_annotations.Book
+        semantic_digital_twin.semantic_annotations.semantic_annotations.Book
     ],
 ):
 
@@ -1784,7 +1784,7 @@ class BookDAO(
 class BookFrontDAO(
     HasBodyDAO,
     DataAccessObject[
-        semantic_digital_twin.adapters.procthor.procthor_semantic_annotations.BookFront
+        semantic_digital_twin.semantic_annotations.semantic_annotations.BookFront
     ],
 ):
 
@@ -1803,7 +1803,7 @@ class BookFrontDAO(
 class ChairDAO(
     HasBodyDAO,
     DataAccessObject[
-        semantic_digital_twin.adapters.procthor.procthor_semantic_annotations.Chair
+        semantic_digital_twin.semantic_annotations.semantic_annotations.Chair
     ],
 ):
 
@@ -1822,7 +1822,7 @@ class ChairDAO(
 class ArmchairDAO(
     ChairDAO,
     DataAccessObject[
-        semantic_digital_twin.adapters.procthor.procthor_semantic_annotations.Armchair
+        semantic_digital_twin.semantic_annotations.semantic_annotations.Armchair
     ],
 ):
 
@@ -1841,7 +1841,7 @@ class ArmchairDAO(
 class OfficeChairDAO(
     ChairDAO,
     DataAccessObject[
-        semantic_digital_twin.adapters.procthor.procthor_semantic_annotations.OfficeChair
+        semantic_digital_twin.semantic_annotations.semantic_annotations.OfficeChair
     ],
 ):
 
@@ -1860,7 +1860,7 @@ class OfficeChairDAO(
 class ClothDAO(
     HasBodyDAO,
     DataAccessObject[
-        semantic_digital_twin.adapters.procthor.procthor_semantic_annotations.Cloth
+        semantic_digital_twin.semantic_annotations.semantic_annotations.Cloth
     ],
 ):
 
@@ -1898,7 +1898,7 @@ class ContainerDAO(
 class BottleDAO(
     ContainerDAO,
     DataAccessObject[
-        semantic_digital_twin.adapters.procthor.procthor_semantic_annotations.Bottle
+        semantic_digital_twin.semantic_annotations.semantic_annotations.Bottle
     ],
 ):
 
@@ -1917,7 +1917,7 @@ class BottleDAO(
 class MustardBottleDAO(
     BottleDAO,
     DataAccessObject[
-        semantic_digital_twin.adapters.procthor.procthor_semantic_annotations.MustardBottle
+        semantic_digital_twin.semantic_annotations.semantic_annotations.MustardBottle
     ],
 ):
 
@@ -1936,7 +1936,7 @@ class MustardBottleDAO(
 class SoapBottleDAO(
     BottleDAO,
     DataAccessObject[
-        semantic_digital_twin.adapters.procthor.procthor_semantic_annotations.SoapBottle
+        semantic_digital_twin.semantic_annotations.semantic_annotations.SoapBottle
     ],
 ):
 
@@ -1955,7 +1955,7 @@ class SoapBottleDAO(
 class WineBottleDAO(
     BottleDAO,
     DataAccessObject[
-        semantic_digital_twin.adapters.procthor.procthor_semantic_annotations.WineBottle
+        semantic_digital_twin.semantic_annotations.semantic_annotations.WineBottle
     ],
 ):
 
@@ -1974,7 +1974,7 @@ class WineBottleDAO(
 class GarbageBinDAO(
     ContainerDAO,
     DataAccessObject[
-        semantic_digital_twin.adapters.procthor.procthor_semantic_annotations.GarbageBin
+        semantic_digital_twin.semantic_annotations.semantic_annotations.GarbageBin
     ],
 ):
 
@@ -1993,7 +1993,7 @@ class GarbageBinDAO(
 class ProcthorBoxDAO(
     ContainerDAO,
     DataAccessObject[
-        semantic_digital_twin.adapters.procthor.procthor_semantic_annotations.ProcthorBox
+        semantic_digital_twin.semantic_annotations.semantic_annotations.ProcthorBox
     ],
 ):
 
@@ -2012,7 +2012,7 @@ class ProcthorBoxDAO(
 class CookingContainerDAO(
     ContainerDAO,
     DataAccessObject[
-        semantic_digital_twin.adapters.procthor.procthor_semantic_annotations.CookingContainer
+        semantic_digital_twin.semantic_annotations.semantic_annotations.CookingContainer
     ],
 ):
 
@@ -2031,7 +2031,7 @@ class CookingContainerDAO(
 class KettleDAO(
     CookingContainerDAO,
     DataAccessObject[
-        semantic_digital_twin.adapters.procthor.procthor_semantic_annotations.Kettle
+        semantic_digital_twin.semantic_annotations.semantic_annotations.Kettle
     ],
 ):
 
@@ -2052,7 +2052,7 @@ class KettleDAO(
 class PanDAO(
     CookingContainerDAO,
     DataAccessObject[
-        semantic_digital_twin.adapters.procthor.procthor_semantic_annotations.Pan
+        semantic_digital_twin.semantic_annotations.semantic_annotations.Pan
     ],
 ):
 
@@ -2073,7 +2073,7 @@ class PanDAO(
 class PotDAO(
     CookingContainerDAO,
     DataAccessObject[
-        semantic_digital_twin.adapters.procthor.procthor_semantic_annotations.Pot
+        semantic_digital_twin.semantic_annotations.semantic_annotations.Pot
     ],
 ):
 
@@ -2094,7 +2094,7 @@ class PotDAO(
 class CuttleryDAO(
     HasBodyDAO,
     DataAccessObject[
-        semantic_digital_twin.adapters.procthor.procthor_semantic_annotations.Cuttlery
+        semantic_digital_twin.semantic_annotations.semantic_annotations.Cuttlery
     ],
 ):
 
@@ -2113,7 +2113,7 @@ class CuttleryDAO(
 class ForkDAO(
     CuttleryDAO,
     DataAccessObject[
-        semantic_digital_twin.adapters.procthor.procthor_semantic_annotations.Fork
+        semantic_digital_twin.semantic_annotations.semantic_annotations.Fork
     ],
 ):
 
@@ -2132,7 +2132,7 @@ class ForkDAO(
 class KnifeDAO(
     CuttleryDAO,
     DataAccessObject[
-        semantic_digital_twin.adapters.procthor.procthor_semantic_annotations.Knife
+        semantic_digital_twin.semantic_annotations.semantic_annotations.Knife
     ],
 ):
 
@@ -2151,7 +2151,7 @@ class KnifeDAO(
 class MilkDAO(
     CuttleryDAO,
     DataAccessObject[
-        semantic_digital_twin.adapters.procthor.procthor_semantic_annotations.Milk
+        semantic_digital_twin.semantic_annotations.semantic_annotations.Milk
     ],
 ):
 
@@ -2174,7 +2174,7 @@ class MilkDAO(
 class SpoonDAO(
     CuttleryDAO,
     DataAccessObject[
-        semantic_digital_twin.adapters.procthor.procthor_semantic_annotations.Spoon
+        semantic_digital_twin.semantic_annotations.semantic_annotations.Spoon
     ],
 ):
 
@@ -2197,7 +2197,7 @@ class SpoonDAO(
 class DecorDAO(
     HasBodyDAO,
     DataAccessObject[
-        semantic_digital_twin.adapters.procthor.procthor_semantic_annotations.Decor
+        semantic_digital_twin.semantic_annotations.semantic_annotations.Decor
     ],
 ):
 
@@ -2216,7 +2216,7 @@ class DecorDAO(
 class WallDecorDAO(
     DecorDAO,
     DataAccessObject[
-        semantic_digital_twin.adapters.procthor.procthor_semantic_annotations.WallDecor
+        semantic_digital_twin.semantic_annotations.semantic_annotations.WallDecor
     ],
 ):
 
@@ -2235,7 +2235,7 @@ class WallDecorDAO(
 class PosterDAO(
     WallDecorDAO,
     DataAccessObject[
-        semantic_digital_twin.adapters.procthor.procthor_semantic_annotations.Poster
+        semantic_digital_twin.semantic_annotations.semantic_annotations.Poster
     ],
 ):
 
@@ -2283,7 +2283,7 @@ class DoorDAO(
 class DrinkingContainerDAO(
     ContainerDAO,
     DataAccessObject[
-        semantic_digital_twin.adapters.procthor.procthor_semantic_annotations.DrinkingContainer
+        semantic_digital_twin.semantic_annotations.semantic_annotations.DrinkingContainer
     ],
 ):
 
@@ -2302,7 +2302,7 @@ class DrinkingContainerDAO(
 class CupDAO(
     DrinkingContainerDAO,
     DataAccessObject[
-        semantic_digital_twin.adapters.procthor.procthor_semantic_annotations.Cup
+        semantic_digital_twin.semantic_annotations.semantic_annotations.Cup
     ],
 ):
 
@@ -2327,7 +2327,7 @@ class CupDAO(
 class MugDAO(
     DrinkingContainerDAO,
     DataAccessObject[
-        semantic_digital_twin.adapters.procthor.procthor_semantic_annotations.Mug
+        semantic_digital_twin.semantic_annotations.semantic_annotations.Mug
     ],
 ):
 
@@ -2348,7 +2348,7 @@ class MugDAO(
 class DroneDAO(
     HasBodyDAO,
     DataAccessObject[
-        semantic_digital_twin.adapters.procthor.procthor_semantic_annotations.Drone
+        semantic_digital_twin.semantic_annotations.semantic_annotations.Drone
     ],
 ):
 
@@ -2367,7 +2367,7 @@ class DroneDAO(
 class FoodDAO(
     HasBodyDAO,
     DataAccessObject[
-        semantic_digital_twin.adapters.procthor.procthor_semantic_annotations.Food
+        semantic_digital_twin.semantic_annotations.semantic_annotations.Food
     ],
 ):
 
@@ -2386,7 +2386,7 @@ class FoodDAO(
 class BreadDAO(
     FoodDAO,
     DataAccessObject[
-        semantic_digital_twin.adapters.procthor.procthor_semantic_annotations.Bread
+        semantic_digital_twin.semantic_annotations.semantic_annotations.Bread
     ],
 ):
 
@@ -2405,7 +2405,7 @@ class BreadDAO(
 class CheezeItDAO(
     FoodDAO,
     DataAccessObject[
-        semantic_digital_twin.adapters.procthor.procthor_semantic_annotations.CheezeIt
+        semantic_digital_twin.semantic_annotations.semantic_annotations.CheezeIt
     ],
 ):
 
@@ -2424,7 +2424,7 @@ class CheezeItDAO(
 class GelatinBoxDAO(
     FoodDAO,
     DataAccessObject[
-        semantic_digital_twin.adapters.procthor.procthor_semantic_annotations.GelatinBox
+        semantic_digital_twin.semantic_annotations.semantic_annotations.GelatinBox
     ],
 ):
 
@@ -2443,7 +2443,7 @@ class GelatinBoxDAO(
 class PringlesDAO(
     FoodDAO,
     DataAccessObject[
-        semantic_digital_twin.adapters.procthor.procthor_semantic_annotations.Pringles
+        semantic_digital_twin.semantic_annotations.semantic_annotations.Pringles
     ],
 ):
 
@@ -2462,7 +2462,7 @@ class PringlesDAO(
 class ProduceDAO(
     FoodDAO,
     DataAccessObject[
-        semantic_digital_twin.adapters.procthor.procthor_semantic_annotations.Produce
+        semantic_digital_twin.semantic_annotations.semantic_annotations.Produce
     ],
 ):
 
@@ -2481,7 +2481,7 @@ class ProduceDAO(
 class AppleDAO(
     ProduceDAO,
     DataAccessObject[
-        semantic_digital_twin.adapters.procthor.procthor_semantic_annotations.Apple
+        semantic_digital_twin.semantic_annotations.semantic_annotations.Apple
     ],
 ):
 
@@ -2500,7 +2500,7 @@ class AppleDAO(
 class BananaDAO(
     ProduceDAO,
     DataAccessObject[
-        semantic_digital_twin.adapters.procthor.procthor_semantic_annotations.Banana
+        semantic_digital_twin.semantic_annotations.semantic_annotations.Banana
     ],
 ):
 
@@ -2519,7 +2519,7 @@ class BananaDAO(
 class LettuceDAO(
     ProduceDAO,
     DataAccessObject[
-        semantic_digital_twin.adapters.procthor.procthor_semantic_annotations.Lettuce
+        semantic_digital_twin.semantic_annotations.semantic_annotations.Lettuce
     ],
 ):
 
@@ -2538,7 +2538,7 @@ class LettuceDAO(
 class OrangeDAO(
     ProduceDAO,
     DataAccessObject[
-        semantic_digital_twin.adapters.procthor.procthor_semantic_annotations.Orange
+        semantic_digital_twin.semantic_annotations.semantic_annotations.Orange
     ],
 ):
 
@@ -2557,7 +2557,7 @@ class OrangeDAO(
 class PotatoDAO(
     ProduceDAO,
     DataAccessObject[
-        semantic_digital_twin.adapters.procthor.procthor_semantic_annotations.Potato
+        semantic_digital_twin.semantic_annotations.semantic_annotations.Potato
     ],
 ):
 
@@ -2576,7 +2576,7 @@ class PotatoDAO(
 class TomatoDAO(
     ProduceDAO,
     DataAccessObject[
-        semantic_digital_twin.adapters.procthor.procthor_semantic_annotations.Tomato
+        semantic_digital_twin.semantic_annotations.semantic_annotations.Tomato
     ],
 ):
 
@@ -2595,7 +2595,7 @@ class TomatoDAO(
 class TomatoSoupDAO(
     FoodDAO,
     DataAccessObject[
-        semantic_digital_twin.adapters.procthor.procthor_semantic_annotations.TomatoSoup
+        semantic_digital_twin.semantic_annotations.semantic_annotations.TomatoSoup
     ],
 ):
 
@@ -2614,7 +2614,7 @@ class TomatoSoupDAO(
 class TunaCanDAO(
     FoodDAO,
     DataAccessObject[
-        semantic_digital_twin.adapters.procthor.procthor_semantic_annotations.TunaCan
+        semantic_digital_twin.semantic_annotations.semantic_annotations.TunaCan
     ],
 ):
 
@@ -2652,7 +2652,7 @@ class HandleDAO(
 class HouseplantDAO(
     HasBodyDAO,
     DataAccessObject[
-        semantic_digital_twin.adapters.procthor.procthor_semantic_annotations.Houseplant
+        semantic_digital_twin.semantic_annotations.semantic_annotations.Houseplant
     ],
 ):
 
@@ -2671,7 +2671,7 @@ class HouseplantDAO(
 class LidDAO(
     HasBodyDAO,
     DataAccessObject[
-        semantic_digital_twin.adapters.procthor.procthor_semantic_annotations.Lid
+        semantic_digital_twin.semantic_annotations.semantic_annotations.Lid
     ],
 ):
 
@@ -2690,7 +2690,7 @@ class LidDAO(
 class PanLidDAO(
     LidDAO,
     DataAccessObject[
-        semantic_digital_twin.adapters.procthor.procthor_semantic_annotations.PanLid
+        semantic_digital_twin.semantic_annotations.semantic_annotations.PanLid
     ],
 ):
 
@@ -2709,7 +2709,7 @@ class PanLidDAO(
 class PotLidDAO(
     LidDAO,
     DataAccessObject[
-        semantic_digital_twin.adapters.procthor.procthor_semantic_annotations.PotLid
+        semantic_digital_twin.semantic_annotations.semantic_annotations.PotLid
     ],
 ):
 
@@ -2728,7 +2728,7 @@ class PotLidDAO(
 class LiquidCapDAO(
     HasBodyDAO,
     DataAccessObject[
-        semantic_digital_twin.adapters.procthor.procthor_semantic_annotations.LiquidCap
+        semantic_digital_twin.semantic_annotations.semantic_annotations.LiquidCap
     ],
 ):
 
@@ -2747,7 +2747,7 @@ class LiquidCapDAO(
 class PenDAO(
     HasBodyDAO,
     DataAccessObject[
-        semantic_digital_twin.adapters.procthor.procthor_semantic_annotations.Pen
+        semantic_digital_twin.semantic_annotations.semantic_annotations.Pen
     ],
 ):
 
@@ -2766,7 +2766,7 @@ class PenDAO(
 class PencilDAO(
     HasBodyDAO,
     DataAccessObject[
-        semantic_digital_twin.adapters.procthor.procthor_semantic_annotations.Pencil
+        semantic_digital_twin.semantic_annotations.semantic_annotations.Pencil
     ],
 ):
 
@@ -2785,7 +2785,7 @@ class PencilDAO(
 class SaltPepperShakerDAO(
     HasBodyDAO,
     DataAccessObject[
-        semantic_digital_twin.adapters.procthor.procthor_semantic_annotations.SaltPepperShaker
+        semantic_digital_twin.semantic_annotations.semantic_annotations.SaltPepperShaker
     ],
 ):
 
@@ -2804,7 +2804,7 @@ class SaltPepperShakerDAO(
 class ShelvingUnitDAO(
     HasBodyDAO,
     DataAccessObject[
-        semantic_digital_twin.adapters.procthor.procthor_semantic_annotations.ShelvingUnit
+        semantic_digital_twin.semantic_annotations.semantic_annotations.ShelvingUnit
     ],
 ):
 
@@ -2823,7 +2823,7 @@ class ShelvingUnitDAO(
 class SinkDAO(
     HasBodyDAO,
     DataAccessObject[
-        semantic_digital_twin.adapters.procthor.procthor_semantic_annotations.Sink
+        semantic_digital_twin.semantic_annotations.semantic_annotations.Sink
     ],
 ):
 
@@ -2842,7 +2842,7 @@ class SinkDAO(
 class SofaDAO(
     HasBodyDAO,
     DataAccessObject[
-        semantic_digital_twin.adapters.procthor.procthor_semantic_annotations.Sofa
+        semantic_digital_twin.semantic_annotations.semantic_annotations.Sofa
     ],
 ):
 
@@ -2861,7 +2861,7 @@ class SofaDAO(
 class SprayBottleDAO(
     HasBodyDAO,
     DataAccessObject[
-        semantic_digital_twin.adapters.procthor.procthor_semantic_annotations.SprayBottle
+        semantic_digital_twin.semantic_annotations.semantic_annotations.SprayBottle
     ],
 ):
 
@@ -2880,7 +2880,7 @@ class SprayBottleDAO(
 class StatueDAO(
     HasBodyDAO,
     DataAccessObject[
-        semantic_digital_twin.adapters.procthor.procthor_semantic_annotations.Statue
+        semantic_digital_twin.semantic_annotations.semantic_annotations.Statue
     ],
 ):
 
@@ -2928,7 +2928,7 @@ class TableDAO(
 class CoffeeTableDAO(
     TableDAO,
     DataAccessObject[
-        semantic_digital_twin.adapters.procthor.procthor_semantic_annotations.CoffeeTable
+        semantic_digital_twin.semantic_annotations.semantic_annotations.CoffeeTable
     ],
 ):
 
@@ -2947,7 +2947,7 @@ class CoffeeTableDAO(
 class DeskDAO(
     TableDAO,
     DataAccessObject[
-        semantic_digital_twin.adapters.procthor.procthor_semantic_annotations.Desk
+        semantic_digital_twin.semantic_annotations.semantic_annotations.Desk
     ],
 ):
 
@@ -2966,7 +2966,7 @@ class DeskDAO(
 class DiningTableDAO(
     TableDAO,
     DataAccessObject[
-        semantic_digital_twin.adapters.procthor.procthor_semantic_annotations.DiningTable
+        semantic_digital_twin.semantic_annotations.semantic_annotations.DiningTable
     ],
 ):
 
@@ -2985,7 +2985,7 @@ class DiningTableDAO(
 class SideTableDAO(
     TableDAO,
     DataAccessObject[
-        semantic_digital_twin.adapters.procthor.procthor_semantic_annotations.SideTable
+        semantic_digital_twin.semantic_annotations.semantic_annotations.SideTable
     ],
 ):
 
@@ -3004,7 +3004,7 @@ class SideTableDAO(
 class VaseDAO(
     HasBodyDAO,
     DataAccessObject[
-        semantic_digital_twin.adapters.procthor.procthor_semantic_annotations.Vase
+        semantic_digital_twin.semantic_annotations.semantic_annotations.Vase
     ],
 ):
 
@@ -3023,7 +3023,7 @@ class VaseDAO(
 class WallPanelDAO(
     HasBodyDAO,
     DataAccessObject[
-        semantic_digital_twin.adapters.procthor.procthor_semantic_annotations.WallPanel
+        semantic_digital_twin.semantic_annotations.semantic_annotations.WallPanel
     ],
 ):
 
@@ -3325,7 +3325,7 @@ class HasSupportingSurfaceDAO(
 class BowlDAO(
     HasBodyDAO,
     DataAccessObject[
-        semantic_digital_twin.adapters.procthor.procthor_semantic_annotations.Bowl
+        semantic_digital_twin.semantic_annotations.semantic_annotations.Bowl
     ],
 ):
 
@@ -3382,7 +3382,7 @@ class FloorDAO(
 class PlateDAO(
     HasBodyDAO,
     DataAccessObject[
-        semantic_digital_twin.adapters.procthor.procthor_semantic_annotations.Plate
+        semantic_digital_twin.semantic_annotations.semantic_annotations.Plate
     ],
 ):
